reorganization of file structure
moved previously unorganized components into their own component folders
