+++ client/src/Components/UI/Dialog/Dialog.module.css
@@ -1,4 +1,4 @@
-@import url(../../constants.css);
+@import url(../../../constants.css);
 
 .dialog {
   display: flex;
