+++ client/src/Components/UI/SubSelectorButton/SubSelectorButton.module.css
@@ -1,4 +1,4 @@
-@import url(../../constants.css);
+@import url(../../../constants.css);
 
 .container {
   display: flex;
